Auth Logic moved to controller and middleware
