+++ routes/auth/userAuth.js
@@ -1,38 +1,9 @@
 var express = require('express');
 var router = express.Router();
-const { User } = require('../../models/index');
-const bcrypt = require('bcrypt')
 
-const { body, validationResult } = require('express-validator');
+const { userAuthValidation, validationHandler } = require('../../middlewares/userAuthValidator');
+const userAuthController = require('../../controllers/userAuthController');
 
-/* GET users listing. */
-router.post('/',
-// username must be an email
-body('username').isEmail(),
-// password must be at least 5 chars long
-body('password').isLength({ min: 5 }),
-
-async (req, res, next) => {
-
-const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }else
-        {
-        const user = await User.findOne({ where: { email: req.body.username } });
-        if (user === null) {
-        res.json({ok: false});
-        } else {
-            bcrypt.compare(req.body.password, user.password, function(err, result) {
-                if (result == true) {
-                    res.send(user);
-                } else{
-                    res.json({ok: false});
-                }
-            });
-    }
-    }
-});
+router.post('/', userAuthValidation, validationHandler, userAuthController.findOne);
 
 module.exports = router;
